Refactor import path for useTimer to use centralized hooks location
